grpc: re-arrange the funcs into different files while maintaining consistency of naming between private/ and public/ services
